requirements.txt
flask
langchain
langchain-xai
flask_cors
markupsafe
Flask-SQLAlchemy
langchain-google-genai
python-dotenv
langgraph
